+++ src/Input.mjs
@@ -1,5 +1,0 @@
-/** @typedef {import("./Option.mjs").Option} Option */
-
-/**
- * @typedef {{disabled?: boolean, "input-mode"?: string, label?: string, max?: string, "max-length"?: number, min?: string, "min-length"?: number, multiple?: boolean, name?: string, options?: Option[], pattern?: string, placeholder?: string, "read-only"?: boolean, required?: boolean, step?: string, subtitle?: string, title?: string, type?: string, value?: string | number | boolean | null}} Input
- */