+++ src/Option.mjs
@@ -1,3 +1,0 @@
-/**
- * @typedef {{disabled?: boolean, label: string, selected?: boolean, title?: string, value: string}} Option
- */